Add files via upload
This project contains a Number Conversion using Java. In which it takes a input from user and according to user's choice it will convert a number.
Number Conversion includes a conversion of decimal number into hexadecimal ,binary as well as octal number and similar for hexadecimal, binary and octal conversion.